py/requirements.txt
fairseq==0.9.0
scipy==1.5.0
scikit-learn==0.23.1
pandas==1.3.4
pyparsing==3.0.6
torch==1.10.0
tqdm==4.62.3
wget==3.2
grpcio-tools==1.29.0
requests